Avoid bug where favorites window wouldn't show max player count on first display
ServerEntryViewModel used a callback to listen for changes to the
server's PlayerCount property. When populating server info, the
PlayerCount was set before the SoftMaxPlayerCount, so when generating
the server info, there wasn't enough info to display the max player
count. Initialize SoftMaxPlayerCount before PlayerCount to get the
expected behaviour the first time the favorites window is rendered.