Client logging is forced on too
This is probably not an issue, since client resets this file on each lunch already. There's no worry that it's going to fill up the users drive.

The only way this is an issue is if you are running your os from an SD card or something where the constant writing may be a problem... most our players dont play from an sd card right?