Fix launcher info failure exception not getting caught properly.
The UrlFallbackSet system raises an AggregateException if it fails, but the code was trying to catch a HttpRequestException. This meant failure to fetch launcher info would hard-crash the launcher entirely.

This isn't going to make people that have this issue able to play the game, but it'll at least make the launcher slightly more robust.