Improve the error message when all configured hubs failed to connect.
I found the "partial error" message to basicly only be useful if you *actually* have multiple hubs setup. Else it's just confusing to the end player... what is a server hub configuration?

Instead, if we get ZERO servers from any of the hubs, we return the regular error. If we get partial results from some hubs, we can use the partial error like before.

Slightly improves the error message to suggest a solution and for this change.