Fix Command-Q not quitting the app under MacOS (#208)
* Fix command-q quitting the app in MacOS

Previously the quit shortcut would do nothing.

This changes the app to be launched using the Avalonia desktop lifetime instead of
manual lifetime management, which lets us process the quit event.

* Move ownership of the running task into the class itself

In theory this gives a cleaner interface for the user, as they no longer need to
keep a reference to the task, or interact with a class property and then
their internal reference.

* Specify an app title

Previously under MacOS the app menu would be just "Avalonia Application".
Unfortunately there’s not currently a place in the launch pipeline that has localisation
available before the app title is set, but for now I figure a fixed English title
is better than the generic one.