Consolidate shared UI into ServerList control (#211)
* Enable dev tools in debug build

* Move the server lists into a single standardised control

Currently both the servers and home tabs contain markup for displaying a server list,
this adds a new control which is used by both to remove the duplication.

NOTE: This currently omits the "Go to the servers tab" button which floats
at the bottom of the favourites list.

* Re-add the "Go to severs tab" button

* Restore missing binding for the server list spinner

* Remove extra binding for list text visibility

User is expected to provide null/empty if text should not be visible.

* Enable "no filter results" message.

* Remove unused property

* Fix filter panel layout

Currently this panel when open will obscure the server list scrollview behind it.

This changes fixes the layout to stack the two panels vertically as you’d expect.

* Add a library which provides hot reloading for XAML layouts.

Enabled only for debug builds.

* Make filter pane resizable

* Dev tools are already enabled by default in debug

This reverts commit 66c7f9e5b9d7d94f0919f4b33813d07ef1dcfb0e.

* Remove unused